+++ src/package/ServiceProvider.php
@@ -4,6 +4,7 @@ namespace PragmaRX\Version\Package;
 
 use Illuminate\Support\Facades\Blade;
 use Illuminate\Support\ServiceProvider as IlluminateServiceProvider;
+use PragmaRX\Version\Package\Console\Commands\Absorb;
 use PragmaRX\Version\Package\Console\Commands\Build;
 use PragmaRX\Version\Package\Console\Commands\Major;
 use PragmaRX\Version\Package\Console\Commands\Minor;
@@ -11,6 +12,8 @@ use PragmaRX\Version\Package\Console\Commands\Patch;
 use PragmaRX\Version\Package\Console\Commands\Refresh;
 use PragmaRX\Version\Package\Console\Commands\Show;
 use PragmaRX\Version\Package\Console\Commands\Version as VersionCommand;
+use PragmaRX\Version\Package\Support\Config;
+use PragmaRX\Yaml\Package\Yaml;
 
 class ServiceProvider extends IlluminateServiceProvider
 {
@@ -21,6 +24,18 @@ class ServiceProvider extends IlluminateServiceProvider
      */
     protected $defer = false;
 
+    /**
+     * The package config.
+     *
+     * @var Config
+     */
+    protected $config;
+
+    /**
+     * Console commands to be instantiated.
+     *
+     * @var array
+     */
     protected $commandList = [
         'pragmarx.version.command' => VersionCommand::class,
 
@@ -35,6 +50,8 @@ class ServiceProvider extends IlluminateServiceProvider
         'pragmarx.version.patch.command' => Patch::class,
 
         'pragmarx.version.refresh.command' => Refresh::class,
+
+        'pragmarx.version.absorb.command' => Absorb::class,
     ];
 
     /**
@@ -43,6 +60,8 @@ class ServiceProvider extends IlluminateServiceProvider
     public function boot()
     {
         $this->publishConfiguration();
+
+        $this->registerBlade();
     }
 
     /**
@@ -66,6 +85,18 @@ class ServiceProvider extends IlluminateServiceProvider
     }
 
     /**
+     * Load config.
+     */
+    protected function loadConfig()
+    {
+        $this->config = new Config(new Yaml());
+
+        $this->config->setConfigFile($this->getConfigFile());
+
+        $this->config->loadConfig();
+    }
+
+    /**
      * Configure config path.
      */
     protected function publishConfiguration()
@@ -84,7 +115,7 @@ class ServiceProvider extends IlluminateServiceProvider
     {
         $this->registerService();
 
-        $this->registerBlade();
+        $this->loadConfig();
 
         $this->registerCommands();
     }
@@ -94,7 +125,7 @@ class ServiceProvider extends IlluminateServiceProvider
      */
     protected function registerBlade()
     {
-        Blade::directive('version', function ($format = Version::DEFAULT_FORMAT) {
+        Blade::directive($this->config->get('blade_directive', 'version'), function ($format = Version::DEFAULT_FORMAT) {
             return "<?php echo app('pragmarx.version')->format($format); ?>";
         });
     }
@@ -130,11 +161,9 @@ class ServiceProvider extends IlluminateServiceProvider
     protected function registerService()
     {
         $this->app->singleton('pragmarx.version', function () {
-            $version = new Version();
+            $version = new Version($this->config);
 
             $version->setConfigFileStub($this->getConfigFileStub());
-
-            $version->loadConfig($this->getConfigFile());
 
             return $version;
         });