+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [0.2.5] - 2018-02-10
+### Added
+- Support for Laravel 5.6 and Symfony 4
+
+## [0.2.4] - 2018-01-09
+### Fixed
+- Package booting Blade too soon
+### Added
+- Blade directive is now configurable
+- Default value for build.length
+
+## [0.2.3] - 2017-12-18
+### Fixed
+- A problem with absorb where git tags were not found
+
+## [0.2.2] - 2017-12-14
+### Added
+- Allow users to suppress app name in version:show Artisan command
+- Ability to absorb current git version and build to the config file
+
+## [0.2.1] - 2017-12-03
+### Fixed
+- Some minor bug fixed
+
 ## [0.2.0] - 2017-12-02
 ### Changed
 - This is a breaking change, in configuration, that's we are moving to 0.2.0